[#138] Move compare income target section to component
