+++ frontend/src/pages/landing/components/index.js
@@ -3,3 +3,4 @@ export { default as ExploreStudies } from "./ExploreStudies";
 export { default as FooterDisclaimer } from "./FooterDisclaimer";
 export { default as Jumbotron } from "./Jumbotron";
 export { default as FrameworkDrivers } from "./FrameworkDrivers";
+export { default as CompareIncomeTarget } from "./CompareIncomeTarget";
