+++ frontend/src/pages/landing/landing.scss
@@ -88,62 +88,4 @@
       }
     }
   }
-
-  .compare-income-target-wrapper {
-    background: #ffffff;
-
-    h2 {
-      color: #00403e;
-      font-family: "RocGrotesk";
-      font-size: 30px;
-      font-style: normal;
-      font-weight: 700;
-      // line-height: 52px;
-    }
-
-    p {
-      color: #005855;
-      font-family: "TabletGothic";
-      font-size: 18px;
-      font-style: normal;
-      font-weight: 400;
-      line-height: 32px;
-    }
-
-    .compare-income-target-info-card-wrapper {
-      padding: 3rem;
-
-      h3 {
-        color: #00302e;
-        font-family: "RocGrotesk";
-        font-size: 20px;
-        font-style: normal;
-        font-weight: 700;
-        line-height: 32px;
-        text-transform: capitalize;
-      }
-
-      p {
-        color: #00302e;
-        font-family: "TabletGothic";
-        font-size: 15px;
-        font-style: normal;
-        font-weight: 400;
-        line-height: 24px;
-      }
-
-      .compare-income-target-card {
-        border-radius: 20px;
-        min-height: 235px;
-
-        &.info-first {
-          background: #47d985;
-        }
-
-        &.info-second {
-          background: #76c8ff;
-        }
-      }
-    }
-  }
 }
